[ chore ] connector 拆出，HashRouter 移到 App
