+++ src/App.tsx
@@ -1,9 +1,10 @@
 import Router from "@/router";
+import { HashRouter } from "react-router-dom";
 
 export default function App() {
   return (
-    <>
+    <HashRouter>
       <Router />
-    </>
+    </HashRouter>
   );
 }
