+++ src/router/index.tsx
@@ -1,14 +1,7 @@
 import React, { useEffect, useState } from "react";
-import { Routes, Route, Navigate, HashRouter } from "react-router-dom";
-import { ConnectedProps, connect } from "react-redux";
+import { Routes, Route, Navigate } from "react-router-dom";
+import { ConnectedProps } from "react-redux";
 import { Layout } from "antd";
-import { changeWorkSpaceAction } from "@/redux/actions/ScreenAction";
-import { getOrganizationsAction } from "@/redux/actions/OrganizationAction";
-import {
-  signInAction,
-  loginAction,
-  loginJwtAction,
-} from "@/redux/actions/AuthAction";
 import { Header } from "@/components/User/Header";
 import { Navbar } from "@/components/User/Navbar";
 import Home from "@/pages/Home";
@@ -22,8 +15,8 @@ import SpinPage from "@/pages/SpinPage";
 import Callback from "@/pages/Login/Callback";
 import WorkSpaceMember from "@/pages/WorkSpace/WorkSpaceMember";
 import WorkSpaceSetting from "@/pages/WorkSpace/WorkSpaceSetting";
-import type { RootState } from "@/redux/store";
 import Invitation from "@/pages/InvitationPage";
+import { connector } from "@/redux/connector";
 
 export type PropsFromRedux = ConnectedProps<typeof connector>;
 
@@ -77,143 +70,124 @@ const AppRouter: React.FC<PropsFromRedux> = (props) => {
   );
 
   return (
-    <HashRouter>
-      <Routes>
-        {load && <Route path="*" element={<SpinPage />} />}
-        {!load && (
-          <>
-            {!login && <Route path="/" element={<Home />}></Route>}
-            <Route
-              path="/invitation/:type/:invitationToken"
-              element={<Invitation />}
-            ></Route>
-            <Route
-              path="/login/:callback"
-              element={
-                <Callback
-                  loginJwt={loginJwt}
-                  getOrganization={getOrganization}
+    <Routes>
+      {load && <Route path="*" element={<SpinPage />} />}
+      {!load && (
+        <>
+          {!login && <Route path="/" element={<Home />}></Route>}
+          <Route
+            path="/invitation/:type/:invitationToken"
+            element={<Invitation />}
+          ></Route>
+          <Route
+            path="/login/:callback"
+            element={
+              <Callback loginJwt={loginJwt} getOrganization={getOrganization} />
+            }
+          ></Route>
+          <Route
+            path="/login"
+            element={
+              <Login
+                signInAction={signInAction}
+                loginAction={loginAction}
+                isSignUpPage={false}
+                getOrganization={getOrganization}
+              />
+            }
+          />
+          <Route
+            path="/signup"
+            element={
+              <Login
+                signInAction={signInAction}
+                loginAction={loginAction}
+                isSignUpPage={true}
+                getOrganization={getOrganization}
+              />
+            }
+          />
+          <Route path="*" element={<ErrorPage />} />
+          {login && (
+            <>
+              {organization?.length ? (
+                <Route
+                  path={"/"}
+                  element={
+                    <Navigate
+                      to={`/workspace/${organization.at(0)?._id}/home`}
+                    />
+                  }
                 />
-              }
-            ></Route>
-            <Route
-              path="/login"
-              element={
-                <Login
-                  signInAction={signInAction}
-                  loginAction={loginAction}
-                  isSignUpPage={false}
-                  getOrganization={getOrganization}
+              ) : (
+                <Route
+                  path={"/"}
+                  element={<Navigate to={"/workspace/new"} />}
                 />
-              }
-            />
-            <Route
-              path="/signup"
-              element={
-                <Login
-                  signInAction={signInAction}
-                  loginAction={loginAction}
-                  isSignUpPage={true}
-                  getOrganization={getOrganization}
-                />
-              }
-            />
-            <Route path="*" element={<ErrorPage />} />
-            {login && (
-              <>
-                {organization?.length ? (
-                  <Route
-                    path={"/"}
-                    element={
-                      <Navigate
-                        to={`/workspace/${organization.at(0)?._id}/home`}
+              )}
+              <Route
+                path={"/workspace/new"}
+                element={<NewWorkSpace getOrganization={getOrganization} />}
+              />
+              <Route
+                index
+                path={`/workspace/:workSpaceId/home`}
+                element={
+                  <LoginLayout
+                    children={
+                      <WorkSpace
+                        setWorkSpace={changeWorkSpace}
+                        getOrganization={getOrganization}
                       />
                     }
                   />
-                ) : (
-                  <Route
-                    path={"/"}
-                    element={<Navigate to={"/workspace/new"} />}
+                }
+              />
+              <Route
+                path={`/workspace/:workSpaceId/members`}
+                element={
+                  <LoginLayout
+                    children={
+                      <WorkSpaceMember
+                        setWorkSpace={changeWorkSpace}
+                        getOrganization={getOrganization}
+                      />
+                    }
                   />
-                )}
-                <Route
-                  path={"/workspace/new"}
-                  element={<NewWorkSpace getOrganization={getOrganization} />}
-                />
-                <Route
-                  index
-                  path={`/workspace/:workSpaceId/home`}
-                  element={
-                    <LoginLayout
-                      children={
-                        <WorkSpace
-                          setWorkSpace={changeWorkSpace}
-                          getOrganization={getOrganization}
-                        />
-                      }
-                    />
-                  }
-                />
-                <Route
-                  path={`/workspace/:workSpaceId/members`}
-                  element={
-                    <LoginLayout
-                      children={
-                        <WorkSpaceMember
-                          setWorkSpace={changeWorkSpace}
-                          getOrganization={getOrganization}
-                        />
-                      }
-                    />
-                  }
-                />
-                <Route
-                  path={`/workspace/:workSpaceId/setting`}
-                  element={
-                    <LoginLayout
-                      children={
-                        <WorkSpaceSetting
-                          setWorkSpace={changeWorkSpace}
-                          getOrganization={getOrganization}
-                        />
-                      }
-                    />
-                  }
-                />
-                <Route
-                  path="/board/:boardId"
-                  element={
-                    <LoginLayout
-                      children={
-                        <Billboard
-                          workSpace={showWorkSpace}
-                          setWorkSpace={changeWorkSpace}
-                        />
-                      }
-                    />
-                  }
-                />
-              </>
-            )}
-          </>
-        )}
-      </Routes>
-    </HashRouter>
+                }
+              />
+              <Route
+                path={`/workspace/:workSpaceId/setting`}
+                element={
+                  <LoginLayout
+                    children={
+                      <WorkSpaceSetting
+                        setWorkSpace={changeWorkSpace}
+                        getOrganization={getOrganization}
+                      />
+                    }
+                  />
+                }
+              />
+              <Route
+                path="/board/:boardId"
+                element={
+                  <LoginLayout
+                    children={
+                      <Billboard
+                        workSpace={showWorkSpace}
+                        setWorkSpace={changeWorkSpace}
+                      />
+                    }
+                  />
+                }
+              />
+            </>
+          )}
+        </>
+      )}
+    </Routes>
   );
 };
-
-const mapStateToProps = (state: RootState) => ({
-  showWorkSpace: state.screen.showWorkSpace,
-  login: state.auth.login,
-  organization: state.user.organization,
-});
-
-const connector = connect(mapStateToProps, {
-  changeWorkSpace: changeWorkSpaceAction,
-  signInAction,
-  loginAction,
-  loginJwt: loginJwtAction,
-  getOrganization: getOrganizationsAction,
-});
 
 export default connector(React.memo(AppRouter));
